Major refactoring under way
allow for flexible data transformations (e.g. using Expressions) and
different "FetchData" modules for different types of data. The tensor
definition is now moved to the DNN class, and the extraction of data
will follow.

+++ SVDModel/SVDCore/core/environmentcell.cpp
@@ -6,9 +6,9 @@ std::vector<std::string> EnvironmentCell::mVariables = {};
 
 void EnvironmentCell::setValue(const int var_idx, double new_value)
 {
-    if (var_idx<0 || var_idx>=mVariables.size())
+    if (var_idx<0 || static_cast<size_t>(var_idx)>=mVariables.size())
         throw std::logic_error("EnvironmentCell::setValue: invalid index.");
     if (mValues.size()<mVariables.size())
         mValues.resize(mVariables.size());
-    mValues[var_idx] = new_value;
+    mValues[static_cast<size_t>(var_idx)] = new_value;
 }
